server/requirements.txt
fastapi
uvicorn[standard]
httpx
orjson
python-dotenv
pydantic
pytest
requests
piper-tts
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
